requirements.txt
python-telegram-bot[job-queue]==20.7
solana==0.30.1
aiohttp==3.9.1
base58==2.1.1
python-dotenv==1.0.0